# backend/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ApproveAllocationInput": {
        "description": "The approval request (FR-08, BR-02, BR-06).\n\n``programmeId`` and ``trainerId`` are **optional consistency checks**, not inputs:\nboth are derived from the prediction. When supplied they must agree with it, which\ncatches a stale screen approving a candidate the operator is no longer looking at.\n\n``weights`` and ``weightsWereSimulated`` are accepted for compatibility with the\nexisting frontend body and then **ignored**. The frozen weights always come from\nthe prediction run itself. A receipt whose weights were supplied by the client\ncould be made to say anything, which would defeat the purpose of freezing them.",
        "properties": {
          "predictionId": {
            "exclusiveMinimum": 0.0,
            "title": "Predictionid",
            "type": "integer"
          },
          "programmeId": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Programmeid"
          },
          "remarks": {
            "default": "",
            "maxLength": 2000,
            "title": "Remarks",
            "type": "string"
          },
          "trainerId": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trainerid"
          },
          "weights": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "deprecated": true,
            "description": "Ignored. Frozen weights come from the run.",
            "title": "Weights"
          },
          "weightsWereSimulated": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "deprecated": true,
            "description": "Ignored. Derived from the run.",
            "title": "Weightsweresimulated"
          }
        },
        "required": [
          "predictionId"
        ],
        "title": "ApproveAllocationInput",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_AllocationListItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AllocationListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "pageSize": {
            "title": "Pagesize",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "totalPages": {
            "title": "Totalpages",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total",
          "totalPages"
        ],
        "title": "Page[AllocationListItem]",
        "type": "object"
      },
...
      "ReportResponse": {
        "description": "A report with the filters that produced it.\n\nThe filters travel with the data so an exported PDF can state its own provenance.\nA report showing figures without saying what it was filtered to is how two people\nend up arguing about numbers that were never comparable.",
        "properties": {
          "chart": {
            "items": {
              "$ref": "#/components/schemas/TrendPoint"
            },
            "title": "Chart",
            "type": "array"
          },
          "filters": {
            "additionalProperties": true,
            "title": "Filters",
            "type": "object"
          },
          "generatedAt": {
            "format": "date-time",
            "title": "Generatedat",
            "type": "string"
          },
          "rowCount": {
            "default": 0,
            "title": "Rowcount",
            "type": "integer"
          },
          "rows": {
            "items": {},
            "title": "Rows",
            "type": "array"
          }
        },
        "required": [
          "rows",
          "generatedAt"
        ],
        "title": "ReportResponse",
        "type": "object"
      },
...
      "TrendPoint": {
        "description": "A labelled point on a line.",
        "properties": {
          "label": {
            "title": "Label",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "label",
          "value"
        ],
        "title": "TrendPoint",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "JWT access token.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "ict@upf.go.ug",
      "name": "Uganda Police Force — Directorate of Human Resource Development"
    },
...
    "license": {
      "name": "Restricted — Uganda Police Force internal use"
    },
    "title": "Trainer Prediction System API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/allocations": {
      "get": {
        "description": "Every approved assignment, filterable by status, programme, trainer, approving officer, and approval date range.",
        "operationId": "list_allocations_api_v1_allocations_get",
        "parameters": [
          {
            "description": "Allocation status.",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Allocation status.",
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "programmeId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Programmeid"
            }
          },
          {
            "in": "query",
            "name": "trainerId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Trainerid"
            }
          },
          {
            "in": "query",
            "name": "approvedBy",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Approvedby"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Pagesize",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sortby"
            }
          },
          {
            "in": "query",
            "name": "sortDir",
            "required": false,
            "schema": {
              "default": "asc",
              "pattern": "^(asc|desc)$",
              "title": "Sortdir",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AllocationListItem_"
                }
              }
            },
            "description": "A page of allocations."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List allocations",
        "tags": [
          "Allocations"
        ]
      },
      "post": {
        "description": "**This is the decision.** BR-02 restricts it to a Training Administrator and BR-06 requires it to be explicit — there is no auto-approve and no bulk approve anywhere in this API.\n\nIn one transaction the server verifies the candidate comes from the current ranking, **re-checks the hard gates against live data** (a trainer may have become unavailable since the run — 409 if so), freezes the score, breakdown, rank, weights and rationale, draws a registry number, sets the allocation `PENDING_TRAINER`, moves the programme to `AWAITING_RESPONSE`, notifies the trainer, and writes the audit entry (BR-07).\n\n`weights` and `weightsWereSimulated` in the body are accepted for compatibility and **ignored** — the frozen weights come from the run itself, because a receipt whose figures were supplied by the client could be made to say anything.",
        "operationId": "approve_api_v1_allocations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveAllocationInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllocationListItem"
                }
              }
            },
            "description": "Approved. The Decision Receipt is the response body."
          },
          "403": {
            "description": "Only a Training Administrator may approve (BR-02)."
          },
          "404": {
            "description": "No such candidate."
          },
          "409": {
            "description": "The ranking is superseded, the requirements changed since it was generated, the programme already has an allocation, or the candidate no longer passes the gates."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approve an allocation (FR-08)",
        "tags": [
          "Allocations"
        ]
      }
    },
...
    "/api/v1/reports/allocation-history": {
      "get": {
        "description": "Who was assigned to what, by whom, and what came of it.",
        "operationId": "allocation_history_api_v1_reports_allocation_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            },
            "description": "Rows, a chart by outcome, and the filters used."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Allocations with their outcomes (FR-11)",
        "tags": [
          "Reports"
        ]
      }
    },
    "/api/v1/reports/performance-trends": {
      "get": {
        "description": "Is delivery getting better or worse? Filterable by category and specialisation.\n\nThese are **raw** means, not the shrunk figures the prediction engine uses. A quarter's mean is a description of what happened; the engine's shrinkage exists to stop a single evaluation dominating a *ranking*, which is a different job.",
        "operationId": "performance_trends_api_v1_reports_performance_trends_get",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "specializationAreaId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Specializationareaid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            },
            "description": "One row per quarter, oldest first."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mean evaluation score by quarter (FR-11)",
        "tags": [
          "Reports"
        ]
      }
    },
    "/api/v1/reports/utilisation": {
      "get": {
        "description": "**This is the report that holds the system to account.** A ranking engine that keeps surfacing the same six names is working exactly as designed and failing at the purpose, and this is where that becomes visible.\n\nTrainers with **no** allocations are included deliberately — an empty row is the finding. A report listing only busy people cannot show who is never used.\n\n`region` filters on the **trainer's posting**, not the course venue: the question is which postings are being drawn on, not where courses ran.",
        "operationId": "utilisation_api_v1_reports_utilisation_get",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "region",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            },
            "description": "Rows, a chart of the busiest ten, and the filters used."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Allocations per trainer (FR-11)",
        "tags": [
          "Reports"
        ]
      }
    },
    "/api/v1/reports/{report_type}/export": {
      "get": {
        "description": "Server-side CSV via `StreamingResponse` — the file is never materialised in memory. `reportType` is `utilisation` or `allocation-history`.\n\nThe export is audited as `REPORT_EXPORTED`, with the filters that produced it.",
        "operationId": "export_report_api_v1_reports__report_type__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "report_type",
            "required": true,
            "schema": {
              "title": "Report Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "csv",
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "A CSV stream."
          },
          "404": {
            "description": "Unknown report type."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export a report as CSV",
        "tags": [
          "Reports"
        ]
      }
    },
...
  },
...
}